requirements.txt
python-jenkins==1.0.2